requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
authlib
python-dotenv
python-jose
httpx
itsdangerous
